+++ src/test/java/org/springframework/test/service/LoveUService.java
@@ -7,4 +7,6 @@ package org.springframework.test.service;
  */
 public interface LoveUService {
   void explode();
+
+  String explodeReturn();
 }