+++ src/test/java/org/springframework/test/service/LoveUServiceWithExceptionImpl.java
@@ -2,17 +2,18 @@ package org.springframework.test.service;
 
 /**
  * @author: DoubleW2w
- * @date: 2024/11/6
+ * @date: 2024/11/9
  * @project: sbs-small-spring
  */
-public class LoveUServiceImpl implements LoveUService {
+public class LoveUServiceWithExceptionImpl implements LoveUService {
   @Override
   public void explode() {
     System.out.println("I Am Missing");
+    throw new RuntimeException();
   }
 
   @Override
   public String explodeReturn() {
-    return "I Love U";
+    return null;
   }
 }